# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Allotjaments Sol",
        "description": "REST API. Projecte Allotjaments Sol. DAW Client i servidor. Per utilitzar les cridades que necesiten token primer\n  s'haura de fer login amb un usuari per obtenir el token i despres utilitzar aquest token a la part on posa Authorize. Per\n  mes informacio, obtenir un usuari o altres peticions utilitzar el link de Contact Equip Sol o mandar un mail a jaumefullana@paucasesnovescifp.cat",
        "contact": {
            "name": "Equip Sol",
            "email": "jaumefullana@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:81/allotjamentssol/public/"
        }
    ],
    "paths": {
...
        "/api/usuari": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Mostrar tots els usuaris.",
                "operationId": "ba97209ecb91dec36b7f480f4dfdd524",
                "responses": {
                    "200": {
                        "description": "Mostrar tots els usuari."
                    },
                    "400": {
                        "description": "Hi ha un error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Modifica un usuari",
                "description": "Modifica un usuari.",
                "operationId": "981ac761a1db7074a73150489d7bd527",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Nom": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Pedro"
                                    },
                                    "Llinatges": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Gimenez Santos"
                                    },
                                    "Contrasenya": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "abcd1234"
                                    },
                                    "CorreuElectronic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "pedro@gmail.com"
                                    },
                                    "DNI": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "12341234Q"
                                    },
                                    "Telefon": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "971123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atribut DNI requerit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Crea un usuari",
                "description": "Crea una nou usuari.",
                "operationId": "1ca7629f432c8b7302834289e1f21f14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Nom": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Pedro"
                                    },
                                    "Llinatges": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Gimenez Santos"
                                    },
                                    "Contrasenya": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "abcd1234"
                                    },
                                    "CorreuElectronic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "pedro@gmail.com"
                                    },
                                    "DNI": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "12341234Q"
                                    },
                                    "Telefon": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "971123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atribut Llinatges requerit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Esborra un usuari",
                "description": "Esborra un usuari.",
                "operationId": "aef5ae7067ae03aa90dbb8d8f1013d24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Tupla no trobada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuari/{id}": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Mostrar un usuari",
                "operationId": "e89a2378971a8080eb1baa3597843274",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de l'usuari",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informació de l'usuari.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Hi ha un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Usuari no trobat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}